+++ src/controllers/favorites.js
