+++ src/routes/cities.js
@@ -2,7 +2,7 @@ var { Router } = require("express");
 var router = Router();
 
 //Destructuring de cities_controller donde tengo la logica
-const { getCitiesApi, getCities } = require("../controllers/controller_cities");
+const { getCitiesApi, getCities } = require("../controllers/cities");
 
 //Get all Cities
 router.get("/cities/Api", getCitiesApi);
