+++ src/routes/interested.js
@@ -1,6 +1,6 @@
 var { Router } = require("express");
 var router = Router();
-const controller = require("../controllers/controller_interested");
+const controller = require("../controllers/interested");
 
 router.post("/interested/userInterested", controller.userInterested);
