+++ src/routes/memberships.js
@@ -4,7 +4,7 @@ var router = Router();
 const {
   createMembership,
   getAllMemebership,
-} = require("../controllers/controller_memberships.js");
+} = require("../controllers/memberships.js");
 
 router.post("/membershisp/createMembership", createMembership);
 router.get("/membershisp/getAllMembership", getAllMemebership);
