+++ src/routes/mercado.js
@@ -1,5 +1,5 @@
 const { Router } = require("express");
-const { premiumController } = require("../controllers/controller_mercado");
+const { premiumController } = require("../controllers/mercado");
 
 const router = Router();
