+++ src/routes/properties.js
@@ -2,7 +2,7 @@ var { Router } = require("express");
 var router = Router();
 
 //Llamado a properties.controller donde tengo la logica
-const controller = require("../controllers/controller_properties.js");
+const controller = require("../controllers/properties.js");
 
 //Create Property
 router.post(
